Fix findAll methond in UserRepository and add data.sql

+++ Assignment_2/src/main/java/demo/UserController.java
@@ -25,6 +25,6 @@ public class UserController {
     @RequestMapping("/users")
     public List<User> getAllUser(@RequestParam(value="page", defaultValue="1") int page,
     		@RequestParam(value="item_per_page", defaultValue="10") int item_per_page) {
-        return this.userRepository.findAll((long)page, (long)item_per_page);
+        return this.userRepository.findAll(page, item_per_page);
     }
 }
